Connection string dialogue:Add name to each connection string
This commit adds Name to the connection string class
so it now contains Name, Connection string and Use managed identity.
The purpose of this is to easier find the correct connection
should one have many.

The list of connection strings is editable.

The rest of the functionality for the dialogue is kept as is.