TopicQueueTree:Set count to number of mess in topic/queue
This commit fixes a bug where we had more in the queue than we fetched.
As the property "MessageCount" was populated both with the total count of messages
and later, when the user chose a queue, the loaded messages
this value was changed and wrong; as the user expects total in queue, not loaded.